pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Whaledetection"
version = "0.0.1"
dependencies = ["seaborn","numpy","librosa","matplotlib","PyWavelets","PyYAML","sounddevice","noisereduce","datasets","huggingface_hub","torch", "pydub"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]